Move doc package to cmd/genji
