+++ cmd/genji/shell/command.go
@@ -12,7 +12,7 @@ import (
 
 	"github.com/genjidb/genji"
 	"github.com/genjidb/genji/cmd/genji/dbutil"
-	"github.com/genjidb/genji/doc"
+	"github.com/genjidb/genji/cmd/genji/doc"
 	"github.com/genjidb/genji/document"
 	errs "github.com/genjidb/genji/errors"
 )
