--- doc/doc_wasm.go
@@ -1,8 +1,0 @@
-// +build wasm
-
-package doc
-
-func DocString(rawExpr string) (string, error) {
-	retrun "stripped from build", nil
-}
-
